Refactor test class layout
