+++ smithy-annotation-processor/src/test/java/software/amazon/smithy/processor/TestProcessorAnnotation.java
@@ -1,11 +1,11 @@
-package software.amazon.smithy.processor.test;
+package software.amazon.smithy.processor;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Target;
 
 @Target(ElementType.PACKAGE)
 public @interface TestProcessorAnnotation {
-    String NAME = "software.amazon.smithy.processor.test.TestProcessorAnnotation";
+    String NAME = "software.amazon.smithy.processor.TestProcessorAnnotation";
 
     String namespace();
 }
