+++ smithy-annotation-processor/src/test/resources/META-INF/services/software.amazon.smithy.build.SmithyBuildPlugin
@@ -1,1 +1,1 @@
-software.amazon.smithy.processor.test.TestBuildPlugin
+software.amazon.smithy.processor.TestBuildPlugin
